Partial implementation of Data.ByteString.Lazy
The crux of this version is this lazy representation:
data ByteString = LPS [P.ByteString]
that is a lazy list of normal strict packed byte string.
There is some redundancy in the representation (zero-length ByteStrings
in the list) so we chose the data type invariant to be that list is either
empty or consists of non-null ByteStrings.

That is:
> invariant (LPS []) = True
> invariant (LPS xs) = L.all (not . P.null) xs

If we don't use this invariant then various functions have several more
tests to do. I think tt keeps things simpler to eliminate the redundancy.
So all functions need to preserve this invariant. Conversion functions
will need to check this (which can be done lazily) or we can provide
unsafe versions where the proof obligation is on the user.

One issue that's not quite clear is what the laziness guarantees should
be (affects things like cons and concat) and should we do anything to
coalase small chunks (particularly affects concat, concatMap etc) since
small chunks are going to generally reduce performance. We probably want
to optimise for the case where chunks are fairly large, eg 4k or bigger.

Several functions are O(n/c) where n is the total length and c is the
average chunk size. If we cannot heuristicly maintain chunk size then
perhaps ew should provide an explicit rechunking function that coalacses
chunks that are below a certain size.