Cleanup of existing Fusion code and wdges of new code for a new method. Rename NoAL -> NoAcc. Remove the noAL value, just use the constructor everywhere. rename type EFL to AccEFL and add several other special cases. Don't parameterise AccEFL by the in and out types, it's always Word8 Since we don't have so many 'Word8's all over the place we don't need the type W = Word8 short hand any more.
Huge new chunk of code for an alternative loop fusion system.
It uses first wrapper elimination to get several loops running in
sequence and allocating just one memory block for them all.
Then within that we fuse adjacent compatible loops (eg up/up, down/down)
But it much more detailed than that because we've got cases for
not just up and down loops but also loops that can run in either
direction (ie they don't use an accumulator) and two important special
cases of that: map and filter. This gives a lot of combinations of
loops that can be fused. 23 to be exact giving rise to 23 fusion RULES.

Currently the fusion is not as reliable as it could be but we believe
that the RRULES matcher in the current version of ghc will deal with it
better. See the analysis in the file for details.