Add QC properties for all v2 rules. Disable down/* */down rules for now
All fused down rules are wrong at the moment, I think due to a bug in
doDownLoop, when filtering. When filtering we end up with a shorter dest
buffer, but shorter from the start, not the end.

When doing down filtering:
    [abcdefg] -> [...defg]

So then, when the wrapper reallocates, it tries as usual to realloc from
the front of the string. Leading to a result:
    [...d]

This doesn't happen when we sequenceLoops down though.

So disable all down fusion for now, until this is resolved. We may need
to return the start offset of the string to loopWrapper, instead of
relying on an offset of 0, when reallocating. 

Added test program, down-fuse.hs, testing the down/down fusion rule,
that reveals this bug (try it in ghci, for example).