Add mallocByteStringWith.
Previously, mallocByteString would just poke a null byte, and return a
buffer. This is bad for cache behaviour, particularly in .Lazy, as we'd
be touching the last cache line, then switching back to the first to
fill it.

This patches introduces mallocByteStringWith:

mallocByteStringWith :: Int -> (Ptr Word8 -> IO Int) -> IO ByteString
mallocByteStringWith l f = do
    fp <- mallocForeignPtrArray (l+1)
    m  <- withForeignPtr fp $ \p -> do
        n <- f p
        poke (p `plusPtr` n) (0::Word8) -- i.e. touch the last cache line last
        return n
    return $ PS fp 0 m

So now we fill the buffer linearly. A side effect is that 'create' is
now:
    create l f = unsafePerformIO $ mallocByteStringWith l (\p -> f p >> return l)

Another benefit is that we need only one 'withForeignPtr' to create a
bytestring, instead of 2.