Add lazy-as-available byte strings
Lazy reading of network sockets is fantastic, but often you need to process
data as it's sent. Currently you can either set the block size to 1 byte or
you have to wait for an entire block to be read.

This adds the (terribly named) hGetSomeContentsN, which returns the contents
of a handle, lazily, but in a timely fashion.