ROLLBACK: hGetContents: use hGet instead of hGetNonBlocking + hWaitForInput + hIsEOF
This patch was based on a misconception about the semantics of
hGetBuf, and caused hGetContents to block more often than it should.
The change to documentation also caused additional confusion:

  http://hackage.haskell.org/trac/ghc/ticket/4041

By rolling back we re-introduce the original bug (GHC bug #3808), but
That bug has a workaround (put the Handle in binary mode).

In the future we will fix the bug properly using hGetBufSome, see

  http://hackage.haskell.org/trac/ghc/ticket/4046