Fix a few incorrect uses of inlinePerformIO The incorrect use of inlinePerformIO resulted in multiple calls to mallocByteString being shared, and hence two different strings sharing the same memory. See http://hackage.haskell.org/trac/ghc/ticket/7270
Consider:
  foo x = s `seq` r
    where
      r = B.map succ x
      s = B.map succ r

The B.map function used a pattern like:
  map f (PS fp s len) = inlinePerformIO $ ...
     fp' <- mallocByteString len

Now, in the foo function above, we have both r and s where the compiler
can see that the 'len' is the same in both cases, and with
inlinePerformIO exposing everything, the compiler is free to share the
two calls to mallocByteString len.

The answer is not to use inlinePerformIO if we are allocating, but to use
unsafeDupablePerformIO instead. Another reminder that inlinePerformIO is
really really unsafe, and that we should be using the ST monad instead.