Avoid copying if possible in `concat`
The `binary` package revealed a rather obvious missing optimization here
when it forced extraneous copies with the evaluation of `concat [a,b]`
where `a` is empty and `b` is large.

Here we rework `Data.ByteString.concat` and `Data.ByteString.Lazy.concat`
to more aggressively avoid unnecessary copies in the face of
concatentations of lists with empty chunks. This rework has the nice
advantage of avoiding allocation during computation of the final buffer
length in the case where a copy is necessary (whereas previously
`checkedSum` would fail to fuse and therefore require allocation for its
list argument).