Avoid infinite loop
primMapByteStringBounded would refuse to write to a buffer that had
just enough space left to fit an item, incorrectly thinking it had a
buffer overflow. Then asking for a new buffer of at least the size to
cover the item could result in getting a buffer of that size, causing
an infinite loop.