Don't perform unaligned writes when forbidden
Some architectures, such as SPARC, trap on unaligned memory accesses,
and poke makes no guarantees about its behaviour if the address is not
sufficiently aligned. Therefore we should use a temporary buffer if
asked to write to an unaligned address, unless we know for sure the
architecture supports it without trapping.