Drop 7.0 and 7.2 and add 8.10 to test matrix
Also dropping "head" which (being 8.7) is no longer worth testing.

GHC 7.0 fails to build QuickCheck 1.13, because it pulls in "text" 1.2.4.0
which in turn fails to build.

GHC 7.2 fails the "hPutBuilder test", because random Unicode Strings
don't roundtrip read/write in GHC 7.2.  Fixing this requires limiting
the range of test inputs when GHC is too old.