Refactor CI avoiding test circular dependencies
Instead of running tests in the parent "bytestring" project, with
circular dependencies from the test frameworks, build and run them in
the child `bytestring-tests` project, which builds the parent as a
dependency that does not run tests, avoiding all the loops.

This requires cabal-install >= 2.4 with cabal.project doing most of the
heavy lifting.