New CI, break test circular dependencies
Instead of running tests in the parent "bytestring" project, with
circular dependencies from the test frameworks, build and run them in
the child `bytestring-tests` project, which builds the parent as a
dependency that does not run tests, avoiding all the loops.

This requires cabal-install >= 2.4 with cabal.project doing most of the
heavy lifting.

This branch prototypes a breaking-change via ekmett's PR#175, and bumps
the version to 0.11.0.0 to make sure that various bounds checks would
normally fail.

And yet, it builds all 8.x targets cleanly, and passes CI, by using
cabal.project to run the test in the bytestring-test child package,
which builds all the requisite dependencies and then runs the tests.

In particular, it even worked with a GHC 8.12 snapshot to test Andrew
Martin's string literal PR in combination with ekmett's PR.

So while this still should not be merged as-is, it is ready for
feedback.  On just the tip commit, which reimagines the CI setup.