Use the built bytestring in tests
The loop-breaking benefit of cabal.project in bytestring-tests derives
from the fact that it pre-builds bytestring (from the parent diretory)
and all its dependencies separately, and only then builds the tests
tests package, on which nothing else depends.

What this means is that we really don't need to and should not build
private copies of the ByteString modules for the tests.  Rather, the
tests just use the pre-build ByteString.

THis assumes that tests don't need access to any non-public ByteString
interfaces, which is the case presently (and presumably going forward).