Compute length at compile time for literal strings (#191)
* Compute length at compile time for literal strings

Add unsafePackLiteral to Data.ByteString.Internal. With GHC-8.10+,
use known-key variant of C `strlen` from `GHC.CString` that supports
constant folding. Also in GHC 8.10, another data constructor of
ForeignPtrContents becomes available: LiteralPtr. For string literals,
this is now used. It saves space when there are lots of literals, and
it improves opportunities for case-of-known data constructor optimizations
when a function scrutinizes the length of a ByteString.

* Fix a few small things. Clean up warnings

* Fix aesthetic issues

* Update GHC version to 8.12

Co-authored-by: Simon Jakobi <simon.jakobi@gmail.com>

* Fix a few small things

* Fix copy-and-paste mistakes

* Use BS constructor in unsafePackAddress

Co-authored-by: Simon Jakobi <simon.jakobi@gmail.com>