Simpler, cleaner faster elemIndices
There's no need for a separate `withForeignPointer` call to subtract the
pointers after `memchr`.  The entire loop is a closure around a fixed
foreign pointer which can safely be reused with just a *single*
`withForeignPointer` in each loop iteration.

In informal measurements, this yields non-trivial performance
improvements, but even without these, the code is simply better.