Don't fill ForeignPtrContents field with bottom
As noted in GHC #17290, ForeignPtrContents really should be strict.
With the addition of the FinalPtr constructor we now have a
sensible value to use in nullForeignPtr other than bottom.

Moreover, another unrelated GHC change (#17760) recently forced
the application of strictness for ForeignPtr's ForeignPtrContents
field. Consequently, we are making this change now. Unfortunately this
means that all earlier versions of bytestring will be unbuilable with
GHC >=9.0.