Cleaned up prop_hgetline_like_s8_hgetline and fixed a Windows testing bug.
On Windows, the test data would be written using the platform newlines.
This means that any lone \n would become a \r\n.
The consequence is that the property would fail to test the
implementation on linux line endings when developing on windows.
The fix is to set the newlineMode to noNewlineTranslation before
writing the test data.