Limit use of unboxed elements from array access
Because of upcoming changes in GHC 9.2, it is best to rebox them before
doing anything with them for compatibility without CPP. In particular,
the invariants we can exploit for avoiding CPP are:

 - The element types in the array reads and array writes matches

 - The element type for the N-width array accesses matches the boxed
   N-width type.

The reboxing done here is extremely local so GHC should have no problem
optimizing it away.

CC @bgamari.

Note this will need a 0.11 backport, but no 0.10 backport. It
cherry-picks without conflicts.