Use `unsafeWithForeignPtr` in `create*` functions.
`withForiegnPtr` can regress performance in upcoming GHC
releases (in particular 9.2.4 and 9.4.1). See the discussion on
https://gitlab.haskell.org/ghc/ghc/-/merge_requests/8597 for more details.

Using `unsafeWithForeignPtr` burdens users of these functions with more
proof obligations about the arguments they provide, as described in the
`continuation-termination` note. These shouldn't be burdensome in the vast
majority of cases, as the only additional restriction is that the continuation
can't unconditionally diverge. However, it might still warrant a major version
bump.